Merging test branch with main branch

+++ includes/matrix.h
@@ -6,7 +6,7 @@
 /*   By: pipolint <pipolint@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/27 15:48:13 by pipolint          #+#    #+#             */
-/*   Updated: 2024/10/09 14:51:30 by pipolint         ###   ########.fr       */
+/*   Updated: 2024/10/23 14:55:33 by pipolint         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,9 @@
 
 # include "structs.h"
 
-t_2dmat	*create_2dmat(float points[2][2]);
-t_3dmat	*create_3dmat(float points[3][3]);
-t_4dmat	*create_4dmat(float points[4][4]);
+t_2dmat	*create_2dmat(double points[2][2]);
+t_3dmat	*create_3dmat(double points[3][3]);
+t_4dmat	*create_4dmat(double points[4][4]);
 void	print_2dmatrix(t_2dmat *m);
 void	print_3dmatrix(t_3dmat *m);
 void	print_4dmatrix(t_4dmat *m);
@@ -27,11 +27,11 @@ t_3dmat	*submat_4d(t_4dmat *matrix, int row, int column);
 t_4dmat	identity();
 t_4dmat	*transpose(t_4dmat *matrix);
 
-float	determinant(t_2dmat *mat_2d, t_3dmat *mat_3d, t_4dmat *mat_4d);
-float	minor_3d(t_3dmat *mat, int row, int column);
-float	cofactor_3d(t_3dmat *mat, int row, int column);
-float	minor_4d(t_4dmat *mat, int row, int column);
-float	cofactor_4d(t_4dmat *mat, int row, int column);
+double	determinant(t_2dmat *mat_2d, t_3dmat *mat_3d, t_4dmat *mat_4d);
+double	minor_3d(t_3dmat *mat, int row, int column);
+double	cofactor_3d(t_3dmat *mat, int row, int column);
+double	minor_4d(t_4dmat *mat, int row, int column);
+double	cofactor_4d(t_4dmat *mat, int row, int column);
 
 t_4dmat	*create_4dcofactor(t_4dmat *mat);
 t_bool	inverse_mat(t_4dmat *mat, t_4dmat **ptr);
