+++ vector_math.c
@@ -6,7 +6,7 @@
 /*   By: pipolint <pipolint@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/17 15:57:58 by pipolint          #+#    #+#             */
-/*   Updated: 2024/10/04 16:43:17 by pipolint         ###   ########.fr       */
+/*   Updated: 2024/10/26 12:09:29 by pipolint         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,14 +45,14 @@ t_tuple	subtract_tuples(t_tuple *vec1, t_tuple *vec2)
 	return (resultant);
 }
 
-void	scalar(t_tuple *vec, float scalar)
+void	scalar(t_tuple *vec, double scalar)
 {
 	vec->x *= scalar;
 	vec->y *= scalar;
 	vec->z *= scalar;
 }
 
-t_tuple	return_scalar(t_tuple *vec, float scalar)
+t_tuple	return_scalar(t_tuple *vec, double scalar)
 {
 	t_tuple	resultant;
 	
@@ -62,7 +62,7 @@ t_tuple	return_scalar(t_tuple *vec, float scalar)
 	return (resultant);
 }
 
-t_tuple	return_at(t_ray *ray, float t)
+t_tuple	return_at(t_ray *ray, double t)
 {
 	t_tuple	new;
